+++ src/main/java/framework/PriceProvider.java
@@ -3,6 +3,6 @@ package framework;
 
 public interface PriceProvider {
 	
-	public int getPriceOfPlayer(int playerId, int year);
-	public Card getExtendedPriceOfPlayer(int playerId, int year);
+	public long getPriceOfPlayer(long playerId, int year);
+	public Card getExtendedPriceOfPlayer(long playerId, int year);
 }
